Fix ga fe

+++ optima/parameters.py
@@ -1189,7 +1189,7 @@ def makesimpars(pars, name=None, keys=None, start=None, end=None, dt=None, tvec=
     if type(keys)==str: keys = [keys] # Listify if string
     if tvec is not None: simpars['tvec'] = tvec
     elif settings is not None: simpars['tvec'] = settings.maketvec(start=start, end=end, dt=dt)
-    else: simpars['tvec'] = linspace(start, end, round((end-start)/dt)+1) # Store time vector with the model parameters -- use linspace rather than arange because Python can't handle floats properly
+    else: simpars['tvec'] = linspace(start, end, int(round((end-start)/dt)+1)) # Store time vector with the model parameters -- use linspace rather than arange because Python can't handle floats properly
     if len(simpars['tvec'])>1: dt = simpars['tvec'][1] - simpars['tvec'][0] # Recalculate dt since must match tvec
     simpars['dt'] = dt  # Store dt
     if smoothness is None: smoothness = int(defaultsmoothness/dt)
